/refactor move LLMProcess class from main.py to core/llm_process.py
